make user service work

+++ src/main/java/codewizards/heal_trip/dataAccess/UserDTO.java
@@ -4,5 +4,5 @@ import codewizards.heal_trip.entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface UserDTO extends JpaRepository<User, Long>{
+public interface UserDTO extends JpaRepository<User, Integer>{
 }
